+++ src/app/commands/coordinape.ts
@@ -1,6 +1,55 @@
-import { CommandContext, SlashCommand, SlashCreator } from 'slash-create';
+import { ButtonStyle, CommandContext, ComponentButton, ComponentType, SlashCommand, SlashCreator } from 'slash-create';
 import { LogUtils } from '../utils/Log';
 import { ServiceSupport } from '../service/ServiceSupport';
+import { CustomId } from '../interactions/customId';
+import { findProfileId } from '@api/findProfileId';
+import { getChannelLinkingStatus } from '@api/getChannelLinkingStatus';
+
+const CONFIGURE_BUTTON: ComponentButton = {
+	type: ComponentType.BUTTON,
+	style: ButtonStyle.SUCCESS,
+	label: 'CONFIGURE',
+	custom_id: CustomId.ConfigButton,
+};
+
+const ASSIGN_BUTTON: ComponentButton = {
+	type: ComponentType.BUTTON,
+	style: ButtonStyle.PRIMARY,
+	label: 'ASSIGN',
+	custom_id: CustomId.AssignButton,
+};
+
+const UNASSIGN_BUTTON: ComponentButton = {
+	type: ComponentType.BUTTON,
+	style: ButtonStyle.DESTRUCTIVE,
+	label: 'UNASSIGN',
+	custom_id: CustomId.UnassignButton,
+};
+
+const UPDATE_ALERTS_BUTTON: ComponentButton = {
+	type: ComponentType.BUTTON,
+	style: ButtonStyle.SUCCESS,
+	label: 'ALERTS',
+	custom_id: CustomId.UpdateAlertsButton,
+};
+
+function getLinkButton({ isLinked }: {isLinked: boolean}): ComponentButton {
+	if (isLinked) {
+		return ({
+			type: ComponentType.BUTTON,
+			style: ButtonStyle.DESTRUCTIVE,
+			label: 'UNLINK',
+			custom_id: CustomId.UnlinkButton,
+		});
+	}
+
+	return ({
+		type: ComponentType.BUTTON,
+		style: ButtonStyle.PRIMARY,
+		label: 'LINK',
+		custom_id: CustomId.LinkButton,
+	});
+}
 
 export default class Coordinape extends SlashCommand {
 	constructor(creator: SlashCreator) {
@@ -22,13 +71,24 @@ export default class Coordinape extends SlashCommand {
 		try {
 			await ctx.defer();
 
-			const { componentActionRows, callbackComponents } = await service.getCallbackComponentsWithRows();
+			const profileId = await findProfileId({ userId: ctx.user.id });
 
-			await ctx.send('Coordinape Single Command', { components: componentActionRows });
-			
-			for (const { component: { custom_id }, callback } of callbackComponents) {
-				ctx.registerComponent(custom_id, callback);
+			const isChannelLinked = await getChannelLinkingStatus({ channelId: ctx.channelID });
+		
+			if (isChannelLinked) {
+				await ctx.send('Coordinape Single Command', { components: [
+					{ type: ComponentType.ACTION_ROW, components: [getLinkButton({ isLinked: !!profileId })] },
+					{ type: ComponentType.ACTION_ROW, components: [ASSIGN_BUTTON, UNASSIGN_BUTTON] },
+					{ type: ComponentType.ACTION_ROW, components: [UPDATE_ALERTS_BUTTON] },
+				] });
+
+				return;
 			}
+
+			await ctx.send('Coordinape Single Command', { components: [
+				{ type: ComponentType.ACTION_ROW, components: [getLinkButton({ isLinked: !!profileId })] },
+				{ type: ComponentType.ACTION_ROW, components: [CONFIGURE_BUTTON] },
+			] });
 		} catch (e) {
 			LogUtils.logError('Welp, something went wrong', e);
 			await service.ephemeralError({ msg: JSON.stringify(e) });
