+++ src/app/interactions/componentInteractions/handlers/common/index.ts
@@ -1,2 +1,3 @@
 export * from './disableAllComponents';
 export * from './disableFirstRowComponentButtons';
+export * from './getAlertsText';
