+++ src/app/interactions/componentInteractions/handlers/configuration/0_handleLinkedCircleAlertsCancel.ts
