+++ src/app/interactions/componentInteractions/handlers/configuration/0_handleUnassignButton.ts
@@ -1,0 +1,27 @@
+import { getCircle } from '@api/getCircle';
+import { ComponentContext, ComponentSelectMenu, ComponentType } from 'slash-create';
+import { CustomId } from 'src/app/interactions/customId';
+import Log from 'src/app/utils/Log';
+import { disableAllComponents } from '../common';
+
+const UNASSIGN_ROLE_USER_SELECT: ComponentSelectMenu = {
+	type: ComponentType.USER_SELECT,
+	placeholder: 'Select the user to remove',
+	custom_id: CustomId.UnassignRoleUserSelect,
+};
+
+export async function handleUnassignButton(ctx: ComponentContext): Promise<void> {
+	try {
+		await ctx.editParent({ components: disableAllComponents(ctx) });
+
+		const { circle } = await getCircle({ channelId: ctx.channelID });
+
+		await ctx.send({
+			content: `Which user would you like to remove from circle ${circle.name}?`,
+			components: [{ type: ComponentType.ACTION_ROW, components: [UNASSIGN_ROLE_USER_SELECT] }],
+		});
+	} catch (error) {
+		await ctx.send({ content: 'An error has occured, please contact coordinape\'s support' });
+		Log.error(error);
+	}
+}
