+++ src/app/interactions/componentInteractions/handlers/configuration/index.ts
@@ -1,10 +1,19 @@
+export * from './0_handleAssignButton';
+export * from './0_handleConfigButton';
+export * from './0_handleConfigButton';
+export * from './0_handleLinkButton';
+export * from './0_handleLinkedCircleAlertsCancel';
+export * from './0_handleLinkedCircleAlertsUpdate';
+export * from './0_handleUnassignButton';
+export * from './0_handleUnlinkButton';
+export * from './0_handleUpdateAlertsButton';
 export * from './1.1_handleCreateNewEntities';
-export * from './2_handleLinkCircles';
 export * from './2.1_handleCircleSelect';
 export * from './2.2_handleRequestApiKeys';
-export * from './3_handleSendAlerts';
+export * from './2_handleLinkCircles';
 export * from './3.1_handleAlertsSelect';
 export * from './3.2_handleAlertsToSend';
 export * from './3.3_handleConfirmAlertsToSend';
 export * from './3.4_handleFrequencyOfAlertsToSend';
+export * from './3_handleSendAlerts';
 export * from './4_handleFinalMessage';
