+++ src/app/interactions/customId.ts
@@ -2,15 +2,20 @@ export enum CustomId {
     AlertsSelect = 'ALERTS_SELECT',
     AlertsSelectConfirmButton = 'ALERTS_SELECT_CONFIRM_BUTTON',
     AllowAlertsButton = 'ALLOW_ALERTS_BUTTON',
+    AssignButton = 'ASSIGN_BUTTON',
     AssignRoleUserSelect = 'ASSIGN_ROLE_USER_SELECT',
     CircleSelect = 'CIRCLE_SELECT',
     CircleSelectNextButton = 'CIRCLE_SELECT_NEXT_BUTTON',
     ConfigButton = 'CONFIGURE_BUTTON',
     ConfigNextButton = 'CONFIG_NEXT_BUTTON',
+    LinkButton = 'LINK_BUTTON',
     LinkCircleButton = 'LINK_CIRCLE_BUTTON',
-    Skip = 'SKIP',
-    UnssignRoleUserSelect = 'UNASSIGN_ROLE_USER_SELECT',
-    LinkedChannelAlertsButton = 'LINKED_CHANNEL_ALERTS_BUTTON',
-    LinkedChannelAlertsUpdateButton = 'LINKED_CHANNEL_ALERTS_UPDATE_BUTTON',
     LinkedChannelAlertsCancelButton = 'LINKED_CHANNEL_ALERTS_CANCEL_BUTTON',
+    LinkedChannelAlertsUpdateButton = 'LINKED_CHANNEL_ALERTS_UPDATE_BUTTON',
+    Skip = 'SKIP',
+    UnassignButton = 'UNASSIGN_BUTTON',
+    UnassignRoleUserSelect = 'UNASSIGN_ROLE_USER_SELECT',
+    UnlinkButton = 'UNLINK_BUTTON',
+    UnssignRoleUserSelect = 'UNASSIGN_ROLE_USER_SELECT',
+    UpdateAlertsButton = 'UPDATE_ALERTS_BUTTON',
 }
