--- src/app/service/components/index.ts
@@ -1,4 +1,0 @@
-export * from './getChangeRoleSelect';
-export * from './getConfigureComponents';
-export * from './getLinkedChannelComponents';
-export * from './getLinkingComponents';
